+++ CHANGELOG.md
@@ -3,9 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 From http://keepachangelog.com
 
-## [Unreleased][unreleased]
+## 3.1.0 - [unreleased]
 ### Added
 - Docs to read the docs :-)
+- Class can be imported more easily as in the example : from debinterface import Interfaces
 
 ### Changed
 - dnsmasqRange.rm_itf_range returns False is nothing was changed
@@ -16,8 +17,6 @@ From http://keepachangelog.com
 - bad if/else
 - failing tests
 - NetworkAdapter : missing pre-down, post-up functions
-
-### Removed
 
 
 ## 3.0.1 - 2016-09-15